+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [3.7.2]
+- `onChangeContentOffset` is deprecated
+- Remove NotificationListener 
+
 ## [3.7.1]
 - Fixed rendering empty and error states in `loadable_list` 
 
@@ -6,7 +10,7 @@
 - Fixed processing of the refresh state in `loadable_list.dart`
 
 ## [3.6.1]
-- Added customisation of scroll controller into lists
+- Added customization of scroll controller into lists
 - Added opportunity to use custom slivers as headers
 
 ## [3.6.0]